reworked inventory
Now inventory is a static length numpy array filled with null items(ID=-1).
Null items can be selected and have a hardness of 0 and cannot be placed.
When a block is picked up it replaces the first null item in the list with itself and when the count of an item hits zero it is replaced with a null item.
We changed all the slots to a sprite group allowing you to click on the hotbar to change the item selected in addition to being able to use the scroll wheel as before.
The full inventory can be opened with 'e'. This creates another 3 rows of slots and stops you from breaking and placing while it is open.
The first Item you click will be highlighted and then the second slot you click will swap with the highlighted one and then no item will be considered selected. This can be done between any 2 slots in hotbar or the greater inventory.
Currently once the inventory fills (all slots have unique items) up you can still break blocks but those blocks become 'void' and are not added to the inventory since it is static length.

Co-Authored-By: Justin 2356115 <101044525+J-HyperNova@users.noreply.github.com>